[#20] Use the OpenGraph Module from Rib
Refactor the OpenGraph Protocol metatag rendering to use the new
Rib.Extra.OpenGraph module.

Change the type of the Message type's _messageAvatarUrl field from Text
to a URI. This requires adding orphan instance of the Aeson type classes
to Zulip.Internal.

Bump the Rib version to the latest master commit.

Closes #20